modifications for pytest
