+++ .idea/python.iml
@@ -4,6 +4,7 @@
     <content url="file://$MODULE_DIR$" />
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="module" module-name="cakeit" />
   </component>
   <component name="TestRunnerService">
     <option name="projectConfiguration" value="py.test" />
